# DomainIntelWeb/openapi.json
{
  "components": {
    "schemas": {
      "ArchiveState": {
        "properties": {
          "archived_to": {
            "title": "Archived To",
            "type": "string"
          }
        },
        "required": [
          "archived_to"
        ],
        "title": "ArchiveState",
        "type": "object"
      },
...
      "CancelState": {
        "properties": {
          "cancelled": {
            "title": "Cancelled",
            "type": "boolean"
          }
        },
        "required": [
          "cancelled"
        ],
        "title": "CancelState",
        "type": "object"
      },
      "CountState": {
        "properties": {
          "deleted": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "boolean"
              }
            ],
            "title": "Deleted"
          }
        },
        "required": [
          "deleted"
        ],
        "title": "CountState",
        "type": "object"
      },
...
      "GenerateRequest": {
        "properties": {
          "action": {
            "enum": [
              "daily",
              "weekly",
              "monthly",
              "quarterly",
              "report",
              "deep_report",
              "impact",
              "lab",
              "bootstrap",
              "coverage",
              "history"
            ],
            "title": "Action",
            "type": "string"
          },
          "event": {
            "default": "",
            "maxLength": 1000,
            "title": "Event",
            "type": "string"
          },
          "kind": {
            "default": "",
            "maxLength": 80,
            "title": "Kind",
            "type": "string"
          },
          "pipeline_mode": {
            "default": "generate",
            "enum": [
              "aggregate",
              "generate"
            ],
            "title": "Pipeline Mode",
            "type": "string"
          },
          "provider": {
            "default": "",
            "maxLength": 80,
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "action"
        ],
        "title": "GenerateRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IndustryMutationState": {
        "properties": {
          "folder": {
            "title": "Folder",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "folder",
          "name"
        ],
        "title": "IndustryMutationState",
        "type": "object"
      },
      "IndustryRename": {
        "properties": {
          "folder": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Folder",
            "type": "string"
          },
          "name": {
            "default": "",
            "maxLength": 120,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "folder"
        ],
        "title": "IndustryRename",
        "type": "object"
      },
...
      "JobAccepted": {
        "properties": {
          "action": {
            "default": "",
            "title": "Action",
            "type": "string"
          },
          "email_delivery": {
            "default": false,
            "title": "Email Delivery",
            "type": "boolean"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "status"
        ],
        "title": "JobAccepted",
        "type": "object"
      },
      "JobOutputState": {
        "properties": {
          "output": {
            "title": "Output",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "output"
        ],
        "title": "JobOutputState",
        "type": "object"
      },
...
      "RestorePreviewState": {
        "properties": {
          "collisions": {
            "items": {
              "type": "string"
            },
            "title": "Collisions",
            "type": "array"
          },
          "folder": {
            "title": "Folder",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "industry",
              "daily"
            ],
            "title": "Kind",
            "type": "string"
          },
          "restorable": {
            "title": "Restorable",
            "type": "boolean"
          },
          "restore_count": {
            "title": "Restore Count",
            "type": "integer"
          },
          "skip_count": {
            "title": "Skip Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "kind",
          "folder",
          "restorable",
          "restore_count",
          "skip_count",
          "collisions"
        ],
        "title": "RestorePreviewState",
        "type": "object"
      },
      "RestoreState": {
        "properties": {
          "folder": {
            "title": "Folder",
            "type": "string"
          },
          "kind": {
            "enum": [
              "industry",
              "daily"
            ],
            "title": "Kind",
            "type": "string"
          },
          "restored": {
            "title": "Restored",
            "type": "integer"
          },
          "skipped": {
            "title": "Skipped",
            "type": "integer"
          }
        },
        "required": [
          "kind",
          "folder",
          "restored",
          "skipped"
        ],
        "title": "RestoreState",
        "type": "object"
      },
...
      "SourceCreate": {
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "note": {
            "default": "",
            "maxLength": 2000,
            "title": "Note",
            "type": "string"
          },
          "publisher_country": {
            "default": "",
            "maxLength": 100,
            "title": "Publisher Country",
            "type": "string"
          },
          "tier": {
            "default": "representative",
            "maxLength": 80,
            "title": "Tier",
            "type": "string"
          },
          "url": {
            "maxLength": 2000,
            "minLength": 8,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "category",
          "name",
          "url"
        ],
        "title": "SourceCreate",
        "type": "object"
      },
      "SourceMutationState": {
        "properties": {
          "added": {
            "title": "Added",
            "type": "boolean"
          }
        },
        "required": [
          "added"
        ],
        "title": "SourceMutationState",
        "type": "object"
      },
      "SourcesState": {
        "properties": {
          "categories": {
            "additionalProperties": {
              "items": {
                "additionalProperties": true,
                "type": "object"
              },
              "type": "array"
            },
            "title": "Categories",
            "type": "object"
          },
          "industry": {
            "title": "Industry",
            "type": "string"
          }
        },
        "required": [
          "industry",
          "categories"
        ],
        "title": "SourcesState",
        "type": "object"
      },
...
      "TrashRestoreRequest": {
        "properties": {
          "desired_folder": {
            "default": "",
            "maxLength": 80,
            "title": "Desired Folder",
            "type": "string"
          }
        },
        "title": "TrashRestoreRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "IntDog Local API",
    "version": "4.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/industries/{folder}": {
      "delete": {
        "operationId": "archive_industry_api_industries__folder__delete",
        "parameters": [
          {
            "in": "path",
            "name": "folder",
            "required": true,
            "schema": {
              "title": "Folder",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArchiveState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Archive Industry",
        "tags": [
          "industries"
        ]
      },
      "patch": {
        "operationId": "rename_industry_api_industries__folder__patch",
        "parameters": [
          {
            "in": "path",
            "name": "folder",
            "required": true,
            "schema": {
              "title": "Folder",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IndustryRename"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndustryMutationState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rename Industry",
        "tags": [
          "industries"
        ]
      }
    },
...
    "/api/industries/{folder}/generate": {
      "post": {
        "operationId": "generate_api_industries__folder__generate_post",
        "parameters": [
          {
            "in": "path",
            "name": "folder",
            "required": true,
            "schema": {
              "title": "Folder",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate",
        "tags": [
          "operations"
        ]
      }
    },
...
    "/api/industries/{folder}/sources": {
      "delete": {
        "operationId": "remove_source_api_industries__folder__sources_delete",
        "parameters": [
          {
            "in": "path",
            "name": "folder",
            "required": true,
            "schema": {
              "title": "Folder",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": true,
            "schema": {
              "title": "Category",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "url",
            "required": true,
            "schema": {
              "title": "Url",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CountState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Source",
        "tags": [
          "sources"
        ]
      },
      "get": {
        "operationId": "sources_api_industries__folder__sources_get",
        "parameters": [
          {
            "in": "path",
            "name": "folder",
            "required": true,
            "schema": {
              "title": "Folder",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourcesState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sources",
        "tags": [
          "sources"
        ]
      },
      "post": {
        "operationId": "add_source_api_industries__folder__sources_post",
        "parameters": [
          {
            "in": "path",
            "name": "folder",
            "required": true,
            "schema": {
              "title": "Folder",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceMutationState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Source",
        "tags": [
          "sources"
        ]
      }
    },
...
    "/api/jobs/{run_id}/cancel": {
      "post": {
        "operationId": "cancel_job_api_jobs__run_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Job",
        "tags": [
          "operations"
        ]
      }
    },
    "/api/jobs/{run_id}/output": {
      "get": {
        "operationId": "job_output_api_jobs__run_id__output_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOutputState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Job Output",
        "tags": [
          "operations"
        ]
      }
    },
    "/api/jobs/{run_id}/retry": {
      "post": {
        "operationId": "retry_job_api_jobs__run_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Job",
        "tags": [
          "operations"
        ]
      }
    },
...
    "/api/trash/{item_id}/preview": {
      "get": {
        "operationId": "preview_api_trash__item_id__preview_get",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestorePreviewState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview",
        "tags": [
          "recovery"
        ]
      }
    },
    "/api/trash/{item_id}/restore": {
      "post": {
        "operationId": "restore_api_trash__item_id__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrashRestoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestoreState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore",
        "tags": [
          "recovery"
        ]
      }
    }
  }
}